cartscreen page bugs fixed, auth component

+++ frontend/src/screens/CartScreen.jsx
@@ -13,16 +13,21 @@ import Loader from '../components/Loader';
 import CartItem from '../components/CartItem';
 
 // -- REDUX RELATED IMPORTS --
-import { getAllCartItems, removeItem } from '../actions/cartActions';
+import { getAllCartItems, qtyReset, removeItem } from '../actions/cartActions';
+import store from '../store';
 
-const CartScreen = ({ match, location, history }) => {
-  const [visibility, setVisibility] = useState(false);
-  const [alertMessage, setMessage] = useState('');
+const CartScreen = ({ history }) => {
+  const [msg, setMsg] = useState({
+    display: false,
+    variant: 'info',
+    message: null,
+  });
+  const [updatedQty, setUpdatedQty] = useState([]);
 
   //redux related
   const dispatch = useDispatch();
   const allCartItems = useSelector((state) => state.cart);
-  const { loading: cartItemsLoading, success, error } = allCartItems;
+  const { loading: cartItemsLoading } = allCartItems;
 
   // variables
   const auth = Auth();
@@ -30,41 +35,84 @@ const CartScreen = ({ match, location, history }) => {
   const calcs = Calculations(cartItems);
   let loading = cartItemsLoading || auth.loading;
 
-  //getting all cart items
   useEffect(() => {
     if (auth.logged === false) {
       history.push('/signin?redirect=cart');
     } else if (auth.logged) dispatch(getAllCartItems());
 
-    return () => axios.CancelToken.source().cancel();
+    const unsubscribe = store.subscribe(() => {
+      let update = store.getState().cart;
+      if (update.success) {
+        setMsgHandler('success');
+      } else if (update.error) {
+        setMsg({ ...msg, message: update.error });
+        setMsgHandler('danger', 3);
+      }
+    });
+
+    return () => {
+      axios.CancelToken.source().cancel();
+      unsubscribe();
+    };
   }, [auth.logged, dispatch]);
 
-  const removeFromCart = async (id, name) => {
-    dispatch(removeItem(id));
-    if (success) {
-      setMessage(`${name} has been removed from your cart`);
-      setVisibility(true);
-      setTimeout(() => {
-        setVisibility(false);
-      }, 2000);
+  const removeFromCart = (id, name) => {
+    if (window.confirm(`Are you sure to delete ${name} from your cart?`))
+      dispatch(removeItem(id));
+  };
+
+  const updatedQtyHandler = (id, qty) => {
+    if (updatedQty.length === 0) {
+      let newUpdatedQty = updatedQty;
+      newUpdatedQty.push({ _id: id, qty });
+      setUpdatedQty(newUpdatedQty);
+    } else {
+      for (let i = 0; i < updatedQty.length; i++) {
+        if (updatedQty[i]._id === id) {
+          let newUpdatedQty = updatedQty;
+          newUpdatedQty[i].qty = qty;
+          setUpdatedQty(newUpdatedQty);
+          break;
+        }
+        if (i === updatedQty.length - 1) {
+          let newUpdatedQty = updatedQty;
+          newUpdatedQty.push({ _id: id, qty });
+          setUpdatedQty(newUpdatedQty);
+        }
+      }
     }
+  };
+
+  const qtyResetHandler = (id, qty) => {
+    dispatch(qtyReset(id, qty));
+    updatedQtyHandler(id, qty);
+    let newMsgObj = msg;
+    newMsgObj.message = 'Updated successfully';
+    setMsg(newMsgObj);
   };
 
   const checkoutHandler = () => {
     history.push('/signin?from=cart&redirect=shipping');
   };
 
+  const setMsgHandler = (variant = 'info', seconds = 2) => {
+    setMsg({ ...msg, display: true, variant });
+    setTimeout(() => {
+      setMsg({ ...msg, display: false });
+    }, seconds * 1000);
+  };
+
   return (
     <>
       {loading ? (
         <Loader />
-      ) : error ? (
-        <Message variant='danger'>{error}</Message>
-      ) : success ? (
+      ) : cartItems ? (
         <Row>
           <Col md={8}>
             <h1>Shopping Cart</h1>
-            {visibility && <Message variant='success'>{alertMessage}</Message>}
+            {msg.display && (
+              <Message variant={msg.variant}>{msg.message}</Message>
+            )}
             {cartItems.length === 0 ? (
               <Message>
                 Your cart is empty <Link to='/'>Go back</Link>
@@ -73,9 +121,10 @@ const CartScreen = ({ match, location, history }) => {
               <ListGroup variant='flush'>
                 {cartItems.map((item, index) => (
                   <CartItem
-                    key={index}
+                    key={item._id}
                     item={item}
-                    dispatch={dispatch}
+                    newQty={updatedQty.find((curr) => curr._id === item._id)}
+                    qtyResetHandler={qtyResetHandler}
                     removeFromCart={removeFromCart}
                   />
                 ))}
